camera responds to size of map

+++ Sokoban Game/Assets/scripts/gameState.cs
@@ -6,8 +6,8 @@ public class GameState : ScriptableObject
     public int moves;
     public int totalMoves;
 
-    public int levelHeight;
-    public int levelWidth;
+    public float levelHeight;
+    public float levelWidth;
     public int boxCount;
     public int goalCount;
